backend/requirements.txt

fastapi>=0.95
uvicorn[standard]>=0.20
sqlalchemy>=1.4
pydantic>=1.10
argon2-cffi>=21.3.0
python-jose>=3.3
email-validator>=1.3
python-dotenv>=1.0
pydantic-settings>=0.2
